\name{Post I Information Functions}
\alias{partial.post.z}
\alias{partial.post.z.1}
\alias{partial.f.mu.s}
\alias{get.cov.param}
\alias{get.cov.post.z}
\alias{get.cov.logit.z}
\alias{get.logor.stat}
\alias{partial.logit.p}
\alias{get.cov.logit.PI}
\title{ Post I Information Functions of EMCluster }
\description{
  All post I information functions are for computing relative quantities and
  can be utilized by advanced developers with caution.

  Currently, these are only for workflows.
}
\references{
  \url{http://maitra.public.iastate.edu/}
}
\author{
  Wei-Chen Chen \email{wccsnow@gmail.com} and Ranjan Maitra.
}
\keyword{post_I_information}
