\name{LMT Functions}
\alias{Gmat}
\alias{partial.logL}
\alias{Iy}
\alias{Iy2}
\alias{postPI}
\alias{GenDataSet}
\alias{GenMixDataSet}
\alias{w.2}
\alias{get.E.chi2}
\alias{get.E.delta}
\alias{pchisq.my}
\alias{gridOne}
\alias{col.ppcontour}
\alias{fill.ppcontour}
\alias{fill.pppoints}
\alias{fill.ppmu}
\alias{q.map.new}
\alias{bh.fdr}
\alias{plotp}
\alias{plotq}
\alias{print.lmt}
\title{ Likelihood Mixture Test (LMT) Functions of EMCluster }
\description{
  All likelihood mixture test (LMT) functions are for testing and can be
  utilized by advanced developers with caution.

  Currently, these are only for workflows.
}
\references{
  \url{http://maitra.public.iastate.edu/}
}
\author{
  Wei-Chen Chen \email{wccsnow@gmail.com}, Ranjan Maitra, and
  Volodymyr Melnykov.
}
\keyword{LMT}
