\name{All Internal Functions}
\alias{plotpt}
\alias{append.BN}
\alias{draw.colors}
\alias{plotlikeprob}
\alias{plotbar}
\alias{plotBN}
\alias{create.colors}
\alias{create.cex}
\alias{match.colcex}
\alias{em.init}
\alias{print.initret}
\alias{em.Model}
\alias{logL.wt}
\alias{dmixmvn.wt}
\alias{emcluster.wt}
\alias{emgroup.wt}
\alias{e.step.wt}
\alias{mb.em.EM}
\alias{mb.rand.EM}
\alias{assign.class.wt}
\alias{meandispersion}
\alias{shortemcluster.wt}
\alias{print.summary.svd}
\alias{print.svd}
\alias{starts.via.svd.wt}
\alias{summary.svd.wt}
\alias{m.step.wt}
\alias{my.format}
\alias{my.cat}
\alias{my.print}
\alias{class2Gamma.wt}
\alias{Gamma2class.wt}
\alias{wt2wot}
\alias{check.dim}
\alias{check.dim.wt}
\alias{getlist}
\alias{print.RRandret}
\alias{print.summary.emret}
\alias{summary.emret.wt}
\alias{color.class}
\title{ All Internal Functions of EMCluster }
\description{
  All internal functions are for testing and can be utilized by
  advanced developers with caution.
}
\references{
  \url{http://maitra.public.iastate.edu/}
}
\author{
  Wei-Chen Chen \email{wccsnow@gmail.com}, Ranjan Maitra, and
  Volodymyr Melnykov.
}
\keyword{internal}
